refactor(faction): replace Faction with Reader struct
- Replaced all usage of Faction struct with new Reader struct for representing faction/group, including in Member, PolityManager, and related logic
- Removed Faction struct and its property drawer from the codebase
- Updated all arrays and fields of type Faction[] to Reader[] in test and nested classes

refactor(editor): update custom drawers and matrix editor

- Replaced PolityDrawer (Faction) with PolityReaderDrawer (Reader) for custom inspector UI
- Registered PolityReaderDrawer for Reader type
- Moved FactionsMatrix editor code to ManagerMatrixEditor.cs with no functional changes
- Removed unused or empty editor scripts

refactor(namespace): unify and clean up namespace usage

- Standardized namespace usage and removed redundant qualifiers throughout the codebase

refactor: consistently use Reader for all faction/group references

- Ensured all runtime and editor code now uses Reader struct for faction/group, improving clarity and maintainability
