--- Assets/Polity Manager/Editor/Manager/Familes.ManagerEditor.cs
@@ -1,20 +1,0 @@
-using UnityEditor;
-using UnityEngine;
-
-namespace Polity
-{
-    public partial class PolityManagerEditor : Editor
-    {
-        // Start is called once before the first execution of Update after the MonoBehaviour is created
-        void Start()
-        {
-
-        }
-
-        // Update is called once per frame
-        void Update()
-        {
-
-        }
-    }
-}
