+++ Assets/Polity Manager/Editor/Manager/ManagerGroupDrawer.cs
