+++ Assets/Polity Manager/Editor/Manager/ManagerMatrixEditor.cs
