--- Assets/Polity Manager/Editor/PolityDrawer.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4e5b02f43ca09354699addcd865b431d
