+++ Assets/Polity Manager/Editor/PolityReaderDrawer.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 014c16bbd6cb20e4ba2b981b63e910df
+guid: 4e5b02f43ca09354699addcd865b431d
